requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
requests==2.31.0
pandas>=1.3.0
numpy>=1.21.0
Pillow==10.0.1

groq>=0.3.0

gtts==2.3.2
opencv-python==4.9.0.80

# RAG / PDF / Vector DB
pdfplumber>=0.10.3
chromadb>=0.5.0
sentence-transformers>=2.6.1

# NOTE: streamlit cloud + webrtc/audio recording often requires extra system deps.
# Keep these optional; the main app supports uploading audio files without them.
# streamlit-webrtc==0.62.1
# av>=10.0.0
# streamlit-mic-recorder
